Initial commit of the react app files!
I added a template that I used for a hackathon project to work off of.
Now we have to remove the hackathon description and implement our own
stuff in it. It has the basic menubar with a menu button that allows
for some simple navigation, which is a good start.

The colors suck though, so we'll have to go back and reconsider the
white-on-white.